feat(comments): Show user avatar instead of app icon
